adicionando validacao de email duplicado

+++ codigo-fonte/login-e-signup/signup.js
@@ -10,25 +10,31 @@ function inserirCadastro() {
 
     db_cadastro = JSON.parse(localStorage.getItem('db_cadastro')) || [];
 
-    let novoId = 1;
-    if (db_cadastro.length > 0) {
-        novoId = db_cadastro[db_cadastro.length - 1].id + 1;
+    var usuarioExistente = db_cadastro.find(cadastro => cadastro.email === email);
+
+    if (usuarioExistente) {
+        alert('Usuário já cadastrado com esse e-mail');
+    } else {
+        let novoId = 1;
+        if (db_cadastro.length > 0) {
+            novoId = db_cadastro[db_cadastro.length - 1].id + 1;
+        }
+
+        var novoCadastro = {
+            id: novoId,
+            nome: nome,
+            sobrenome: sobrenome,
+            email: email,
+            senha: senha,
+        }
+
+        db_cadastro.push(novoCadastro);
+        localStorage.setItem('db_cadastro', JSON.stringify(db_cadastro));
+
+        alert('Cadastro realizado com sucesso!');
+
+        window.location.href = "../home/index.html";
     }
-
-    var novoCadastro = {
-        id: novoId,
-        nome: nome,
-        sobrenome: sobrenome,
-        email: email,
-        senha: senha,
-    }
-
-    db_cadastro.push(novoCadastro);
-    localStorage.setItem('db_cadastro', JSON.stringify(db_cadastro));
-
-    alert('Cadastro realizado com sucesso!');
-
-    window.location.href = "../home/index.html";
 }
 
 document.addEventListener('DOMContentLoaded', initPage);
